Add Row Bandit skeleton code
At the moment, the bandit makes no actual decisions and is actually just
an open row policy in disguise. However, all the `wiring` is there, so
we need only modify row_bandit.cc and row_bandit.h in the future.

The bandit implementation is from
https://github.com/nicoleorzan/Multi-armed-bandit-RL
It's very simple and immature, maybe we'll want to swap it out.